Move workflow executions to components
