+++ frontend/src/components/executions/event-details.tsx
